requirements.txt
asyncpg==0.29.0
email_validator==2.2.0
fastapi==0.112.1
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
pillow==10.4.0
pydantic==2.8.2
pyjwt==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
uvicorn==0.30.6
werkzeug==3.0.3